Another 2x performance improvement on Zen 4.

Before:
```
verify                  time:   [204.92 ms 205.10 ms 205.30 ms]
```
After:
```
verify                  time:   [102.38 ms 102.43 ms 102.53 ms]
```

On this particular CPU verification is 16x faster than proving, which matches expected dual AES units per CPU core with 0.5 clocks per instruction:
```
2 AES units * 4 blocks at a time / 0.5 CPI = 16
```

:muscle: